In our GSDK copy of ot-br-posix ("super/util/third_party/ot-br-posix"), we don't include a copy of the stack under its submodule structure "third_party/openthread/repo".
Rather we sym-link it to the stack that we ship, which is under "super/util/third_party/openthread"

When building using Docker with our GSDK (as we are currently prototyping for Jenkins continuous integration), we need to stage the working directory so that the sym-linked directory is also included, which means (at least) from "super/util/third_party".  We correspondingly need to copy both directories to our Docker staging area.