Implement RCP error recovery on default OTBR with 100 retries, similar to multiprotocol.
We find that this alleviates unexpected shutdowns on OTBR due to serial communication issues found during OTBR certification tests.  The root cause of the actual error has not been found as the problem is intermittent, but we're going to treat this as a workaround, and recommended band-aid for RCP recovery.