[script] always create `sysctl.conf` in `rt_tables_install()`
When using `docker-rpi-emu` to customize a Raspberry Pi OS image with OTBR pre-installed,
the resultant image does not include the `/etc/sysctl.conf` file.

I understand that the original intention of the `if without DOCKER` statement was probably
to get around the problem of the net.core.optmem_max` kernal parameter not being present in
a docker environment.

This fix will ensure `/etc/sysctl.conf` is still created so it can be loaded when the
customized RPi OS image is running on an physical RPi.